type error fixed

+++ CookieData/Entities/GameAccount.cs
@@ -5,7 +5,7 @@ namespace CookieData.Entities
     public class GameAccount : BaseEntity
     {
         public double Networth { get; set; }
-        public long Cookies { get; set; }
+        public double Cookies { get; set; }
         public long Clicks { get; set; }
         public ClickUpgrade ClickUpgrade { get; set; }
         public IEnumerable<Upgrade> Upgrades { get; set; }
